refact: extract withCallViewModel in test utils file
